img_sim/requirements.txt
pandas
pillow
numpy>=1.21.0,<2.0.0
sentence-transformers
faiss-gpu
torch
torchvision
torchaudio
fastapi
uvicorn
